+++ BaseUI/src/components/avatar/Avatar.tsx
@@ -1,16 +1,9 @@
 /* eslint-disable react-native/no-inline-styles */
 import React, {Component} from 'react';
-
-import {ImageBackground, StyleSheet, TouchableOpacity} from 'react-native';
+import {ImageBackground, TouchableOpacity} from 'react-native';
 import FontAwesome from 'react-native-vector-icons/FontAwesome';
-
-interface IAvatarProps {
-  avatarStyle?: object;
-  avatarImg: object;
-  onPress(): void;
-  disabled: boolean;
-  isEditMode: boolean;
-}
+import styles from './avatarStyles';
+import {IAvatarProps} from '../../helpers/ts-helpers/interfaces';
 
 class Avatar extends Component<IAvatarProps, {}> {
   render() {
@@ -18,7 +11,7 @@ class Avatar extends Component<IAvatarProps, {}> {
       <TouchableOpacity
         activeOpacity={0.8}
         onPress={this.props.onPress}
-        disabled={this.props.disabled}
+        disabled={!this.props.isEditMode}
         style={this.props.avatarStyle}>
         <ImageBackground
           style={styles.imageBackgroud}
@@ -38,17 +31,17 @@ class Avatar extends Component<IAvatarProps, {}> {
   }
 }
 
-const styles = StyleSheet.create({
-  imageBackgroud: {
-    width: 80,
-    height: 80,
-  },
-  cameraIcon: {
-    position: 'absolute',
-    right: 7,
-    bottom: 5,
-    opacity: 0.5,
-  },
-});
+// const styles = StyleSheet.create({
+//   imageBackgroud: {
+//     width: 80,
+//     height: 80,
+//   },
+//   cameraIcon: {
+//     position: 'absolute',
+//     right: 7,
+//     bottom: 5,
+//     opacity: 0.5,
+//   },
+// });
 
 export default Avatar;
