+++ BaseUI/src/components/avatar/avatarStyles.ts
@@ -1,0 +1,16 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+  imageBackgroud: {
+    width: 80,
+    height: 80,
+  },
+  cameraIcon: {
+    position: 'absolute',
+    right: 7,
+    bottom: 5,
+    opacity: 0.5,
+  },
+});
+
+export default styles;
