+++ BaseUI/src/components/backgroudForm/backgroundFormStyles.ts
@@ -1,0 +1,19 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+  backgroundImage: {
+    flex: 1,
+    justifyContent: 'flex-end',
+  },
+  viewStyle: {
+    paddingTop: 30,
+    paddingBottom: 50,
+    paddingHorizontal: 20,
+    width: '100%',
+    backgroundColor: '#ffffff',
+    borderTopLeftRadius: 20,
+    borderTopRightRadius: 20,
+  },
+});
+
+export default styles;
