+++ BaseUI/src/components/credentialTextInput/CredentialTextInput.tsx
@@ -1,17 +1,7 @@
 import React, {Component} from 'react';
-
-import {TextInput, Text, StyleSheet, View} from 'react-native';
-
-interface ICredentialTextInputProps {
-  placeholder: string;
-  secureTextEntry?: boolean;
-  inputStyle?: object;
-  value?: string;
-  editable?: boolean;
-  onFocus?(): void;
-  onChangeText?(text: string): void;
-  errorMessage?: string;
-}
+import {TextInput, Text, View} from 'react-native';
+import styles from './credentialTextInputStyles';
+import {ICredentialTextInputProps} from '../../helpers/ts-helpers/interfaces';
 
 class CredentialTextInput extends Component<ICredentialTextInputProps, {}> {
   render() {
@@ -36,23 +26,5 @@ class CredentialTextInput extends Component<ICredentialTextInputProps, {}> {
     );
   }
 }
-
-const styles = StyleSheet.create({
-  textWrapper: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-  },
-  errorMsg: {
-    fontSize: 10,
-    color: '#ff0000',
-  },
-  textInputStyle: {
-    paddingVertical: 8,
-    width: '100%',
-    borderBottomWidth: 1,
-    borderColor: 'rgb(181,182,221)',
-    fontSize: 15,
-  },
-});
 
 export default CredentialTextInput;
