+++ BaseUI/src/components/filledButton/filledButtonStyles.ts
@@ -1,0 +1,18 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+  button: {
+    justifyContent: 'center',
+    alignItems: 'center',
+    width: '100%',
+    height: 50,
+    borderRadius: 25,
+    backgroundColor: 'rgb(64, 80, 164)',
+  },
+  text: {
+    fontSize: 15,
+    color: '#ffffff',
+  },
+});
+
+export default styles;
