+++ BaseUI/src/components/socialNetworkButton/socialNetworkButtonStyles.ts
@@ -1,0 +1,14 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+  touchableOpacityStyle: {
+    justifyContent: 'center',
+    alignItems: 'center',
+    width: 45,
+    height: 45,
+    borderRadius: 8,
+    backgroundColor: 'rgb(255, 103, 97)',
+  },
+});
+
+export default styles;
