+++ BaseUI/src/components/socialNetworkButtonsForm/socialNetworkButtonsFormStyles.ts
@@ -1,0 +1,13 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+  containerStyle: {
+    flexDirection: 'row',
+    justifyContent: 'center',
+  },
+  buttonStyle: {
+    marginRight: 11,
+  },
+});
+
+export default styles;
