+++ BaseUI/src/components/textButton/textButtonStyles.ts
@@ -1,0 +1,10 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+  touchableOpacityStyle: {},
+  textStyle: {
+    fontSize: 13,
+  },
+});
+
+export default styles;
