+++ BaseUI/src/helpers/ts-helpers/enums.ts
@@ -1,0 +1,3 @@
+export enum KeyStorage {
+  UserProfile = '@userProfile',
+}
