+++ BaseUI/src/screens/loginScreen/loginScreenStyles.ts
@@ -1,0 +1,34 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+  viewStyle: {
+    paddingTop: 30,
+    paddingBottom: 50,
+    paddingHorizontal: 20,
+    width: '100%',
+    backgroundColor: '#ffffff',
+    borderTopLeftRadius: 20,
+    borderTopRightRadius: 20,
+  },
+  inputContainer: {
+    marginBottom: 30,
+  },
+  inputStyle: {
+    marginBottom: 15,
+  },
+  textButtonStyle: {
+    marginBottom: 40,
+  },
+  orSignWithStyle: {
+    marginTop: 18,
+    marginBottom: 18,
+    textAlign: 'center',
+    fontSize: 13,
+    color: 'rgb(181, 182, 221)',
+  },
+  buttonsContainerStyle: {
+    marginBottom: 30,
+  },
+});
+
+export default styles;
